requirements.txt
Django~=4.2.2
psycopg2-binary~=2.9.6
django-environ~=0.10.0
djangorestframework~=3.14.0
social-auth-app-django~=5.2.0
django-filter~=23.2
